requirements.txt
questionary==2.1.0
discord.py-self==2.0.0
rich==13.9.4
psutil==7.0.0
pypresence==4.3.0
colorama==0.4.6
inquirer==3.4.0
requests==2.32.3
art==6.4
datetime==5.5